feat(TypeHandlerLibrary): move most part of code from TypeHandlerLibraryIml
